Removing nexus wrapper, insturment, component factory
