+++ tests/test_component_factory.py
@@ -5,7 +5,6 @@ import numpy as np
 from nexus_constructor.component.chopper_shape import ChopperShape
 from nexus_constructor.component.component_shape import ComponentShape
 from nexus_constructor.component.pixel_shape import PixelShape
-from nexus_constructor.component.component_factory import create_component
 from nexus_constructor.model.geometry import OFFGeometryNoNexus, OFFGeometryNexus
 from nexus_constructor.model.geometry import NoShapeGeometry
